Changed one-hot encoding to code as 0/1 not true/false

+++ src/data/encoders.py
@@ -19,7 +19,7 @@ def encode_categorical(df: pd.DataFrame) -> pd.DataFrame:
 
 
 def encode_one_hot(df: pd.DataFrame) -> pd.DataFrame:
-    return pd.get_dummies(df, prefix=df.columns, columns=df.columns)
+    return pd.get_dummies(df, prefix=df.columns, columns=df.columns, dtype=int)
 
 
 class CatEncodingStrategy(Enum):
